--- core/src/main/resources/application.conf
@@ -1,3 +1,0 @@
-//odinson.dataDir = "/D%3A/Users/kwa/Documents/MyData/Projects/odinson-tests-project/corpora/100"
-//odinson.docDir = "/D%3A/Users/kwa/Documents/MyData/Projects/odinson-tests-project/corpora/100/docs"
-//odinson.indexDir = "/D:/Users/kwa/Documents/MyData/Projects/odinson-tests-project/corpora/100/index"
