+++ core/src/main/resources/reference.conf
@@ -24,7 +24,7 @@ odinson {
   computeTotalHits = true
 
   state {
-    provider = "fastsql" // "fastsql" "sql" "file" "memory"
+    provider = "memory" // "fastsql" "file" "memory" "mock" "sql"
     fastsql {
       // See http://www.h2database.com/html/features.html.
       url = "jdbc:h2:mem:odinson;trace_level_file=0" // memory
