+++ core/src/main/resources/timerapp.conf
@@ -1,0 +1,3 @@
+odinson.dataDir  = "/D:/Users/kwa/Documents/MyData/Projects/odinson-tests-project/corpora/100"
+odinson.docDir   = "/D:/Users/kwa/Documents/MyData/Projects/odinson-tests-project/corpora/100/docs"
+odinson.indexDir = "/D:/Users/kwa/Documents/MyData/Projects/odinson-tests-project/corpora/100/index"
